+++ example/lib/enemy/goblin.dart
@@ -7,7 +7,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class Goblin extends SimpleEnemy {
-  double attack = 25;
+  double attack = 20;
   bool _seePlayerClose = false;
 
   Goblin(Position initPosition)